% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs.R
\docType{data}
\name{bugs}
\alias{bugs}
\title{Bacterial growth in rat mesenteric lymph}
\format{A dataframe with 200 rows and 8 variables:
\describe{
  \item{Rat}{ID number from 1-20}
  \item{State}{Disease state under which lymph was collected: Normal (Control) and three disease conditioned states
  (AP, DM and HS)}
  \item{Bacteria}{Bacterial species (Ec, Ef, Kp, Pa and Sa)}
  \item{Time}{4 and 24 hours post-inoculation}
  \item{Baseline}{Concentration of the initial bacterial inoculation}
  \item{Cells}{Concentraton of bacteria}
}}
\source{
Unpublished data.
}
\usage{
bugs
}
\description{
These data are from an experiment conducted to investigate the growth of five bacterial strains over time (4 and 24 hours post-inoculation) in mesenteric lymph harvested from normal and diseased rats.
}
\keyword{datasets}
