% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintplot.R
\name{confintplot}
\alias{confintplot}
\title{A ggplot2 wrapper function for generating plots with confidence interval error bars}
\usage{
confintplot(data, y, x, pointSize = 3.5, showSignif = FALSE, pvalVar,
  legendShapeSize = 3, legendTextSize = 10, legendTitle, zeroline = TRUE,
  se, df, alpha, facets, xLabel, yLabel, axisLabelSize = 14,
  tickMarkLabelSize = 12, xTickLabelAngle = 0, xStripTextSize = 12,
  yStripTextSize = 12)
}
\arguments{
\item{data}{a dataframe, assumed to be generated by \link{makeSummaryTable}.}

\item{y}{name of column in \code{data} containing predicted values, usually predicted means
or pairwise differences between predicted means.}

\item{x}{name of variable or factor in \code{data} to be plotted on x-axis.}

\item{pointSize}{size of plotting character for predicted values.}

\item{showSignif}{logical (default \code{FALSE}). Controls appearance of plotting characters
used for predicted values in \code{y}. If \code{TRUE}, a value must be supplied to
\code{alpha} argument and, in this case, a filled circle is drawn for predicted values with a
corresponding p-value less than \code{alpha} and an open circle otherwise.}

\item{pvalVar}{character string giving column name in \code{data} containing p-values}

\item{legendShapeSize}{size of plotting characters in legend}

\item{legendTextSize}{font size of legend text}

\item{legendTitle}{character string, used in title of legend}

\item{zeroline}{logical (default \code{TRUE}), draws a dashed horizontal line on the plot at
\code{y = 0}}

\item{se}{name of column in \code{data} containing the standard errors of the predicted values
provided to \code{y}. A value must be supplied when \code{data} does not include an LSD column.}

\item{df}{degrees of freedom with which the standard error (see \code{se}) was estimated.
A value must be supplied when \code{data} does not include an LSD column.}

\item{alpha}{level of significance needed to calculate confidence limits. A value must be
supplied when \code{data} does not include an LSD column.}

\item{facets}{character vector of factor names whose levels define a matrix of row and/or column
panels in the plot. This vector must be of length 2. Thus, if only one facet factor is needed,
fill the vector with the character string \code{"."} (see Examples below). The order of the
factor names, one of which may be the character string \code{"."}, determine whether the levels
are displayed in rows or columns, with first element (factor name) corresponding to row panels
and the second to column panels.}

\item{xLabel}{x-axis label (character string)}

\item{yLabel}{y-axis label (character string)}

\item{axisLabelSize}{font size of x- and y-axis labels. Currently, these cannot be set
independently of one another.}

\item{tickMarkLabelSize}{font size of x- and y-axis tick mark labels. Currently, these cannot be
set independently of one another.}

\item{xTickLabelAngle}{angle (0, 45 or 90; default 0) of x-axis tick labels}

\item{xStripTextSize}{x strip text size in pts}

\item{yStripTextSize}{y strip text size in pts}
}
\value{
Generates one- (default) and two-way boxplots. A one-way boxplot is generated when
\code{facets} are not supplied (default). A two-way boxplot is generated when \code{facets}
are supplied.

Plot aesthetics, themes, etc. may be saved in an R object so that other
layers may be added.
}
\description{
This function generates plots of predicted values (e.g. means, differences between means, etc)
with their \eqn{100(1 - \alpha)}\% confidence intervals, where \eqn{\alpha} denotes the level
of significance.
}
\details{
This function is a wrapper to generate side-by-side boxplots for either a single factor
or a pair of factors.
}
\examples{
library(ggplot2)
# Means with 95\% CI error bars
data(crdData)
crd.lm <- lm(logAUC ~ Surgery, data = crdData)
crd.pm <- predictmeans(crd.lm, modelterm = "Surgery", pairwise = TRUE, plot=FALSE)
crd.tab <- makeSummaryTable(crd.pm)
confintplot(data=crd.tab, y="Difference", x="Comparison")

# use open and closed circles for p-values less than 0.05
confintplot(data=crd.tab, y="Difference", x="Comparison",
            showSignif=TRUE, pvalVar="p", alpha = 0.05)

# control legend plotting character (shape) and text sizes
confintplot(data=crd.tab, y="Difference", x="Comparison",
            showSignif=TRUE, pvalVar="p", alpha = 0.05,
            legendShapeSize = 5, legendTextSize = 14)
}
\seealso{
\code{\link[ggplot2]{ggplot}} \code{\link[ggplot2]{geom_point}}

\code{\link[ggplot2]{geom_errorbar}} \code{\link[ggplot2]{aes_string}}

\code{\link[ggplot2]{facet_wrap}}
}
\author{
Katya Ruggiero
}
