% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crdData.R
\docType{data}
\name{crdData}
\alias{crdData}
\title{Simulated data set for a completely randomised design (CRD)}
\format{A data frame with 12 rows and 3 variables:
\describe{
  \item{Surgery}{type of surgical conditioning}
  \item{Rat}{animal IDs}
  \item{logAUC}{log-transformed area under the curve, a proxy for log-abundance of protein}
}}
\usage{
crdData
}
\description{
A dataset containing the log-transformed protein abundances
in 12 rats randomly assigned to one of three surgical conditionings.
}
\keyword{datasets}
