% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSmallestMean.R
\name{findSmallestMean}
\alias{findSmallestMean}
\title{Finds lowest value among predicted means of a set of treatment groups.}
\usage{
findSmallestMean(data, y, eqLevelFactors)
}
\arguments{
\item{data}{dataframe containing columns which are factors and one column containing
the predicted means for each factorial treatment combination (defined by the values in each
row across the set of factors in the dataframe).}

\item{y}{character string of column name in \code{data} containing means.}

\item{eqLevelFactors}{character vector of columns names in \code{data}
corresponding to those factors whose levels are constant (i.e. have the same value) in the
pair of factorial treatment combinations being compared.}
}
\value{
A dataframe comprising a column for each factor in \code{eqLevelFactors} and a column
for the minimum mean for each factorial treatment combination given by the levels of the factors
in \code{eqLevelFactors}.
}
\description{
The function was primarily created to be used (called) by \link{makeLSDlist} to create
one of the two dataframes needed by \link{meansLSDplot} to generate plots of means with LSD
bars.
}
\examples{
library(nlme)
library(predictmeans)
Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)
# library(lme4)
# fm <- lmer(yield ~ nitro*Variety+(1|Block/Variety), data=Oats)
pm <- predictmeans(fm, "nitro:Variety", pairwise=TRUE, plot=FALSE)

# generate dataframe of means for each factorial treatment combination
predMeans.pm <- means2df(pm)
findSmallestMean(predMeans.pm, y = "Mean", eqLevelFactors = "Variety")
}
\author{
Katya Ruggiero
}
