% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFixedLevelComparisons.R
\name{getFixedLevelComparisons}
\alias{getFixedLevelComparisons}
\title{Finds rows of post-hoc tests summary table in which pairs of treatment groups differ in the
levels of only one factor}
\usage{
getFixedLevelComparisons(comparisonNames, sepChar)
}
\arguments{
\item{comparisonNames}{a vector of character strings containing the names of the
treatment combinations being contrasted in each pairwise comparison of means. The first column of
the post-hoc tests summary table generated by \code{\link{makeSummaryTable}} produces the
type of input required.}

\item{sepChar}{the character used to concatenate the factor levels in each treatment
combination in \code{comparisonNames}.}
}
\value{
a numeric vector of the rows of interest in the post-hoc tests summary table.
}
\description{
This function identifies the rows in a post-hoc tests summary table (a dataframe object),
generated by \code{\link{makeSummaryTable}}, that correspond to comparisons between pairs of factorial
treatment combinations in which all but one of the levels are equal. Put another way, the two
factorial treatment combinations differ in the levels of only one (and the same) factor in each
treatment.
}
\details{
The row numbers returned by this function
may be used to extract a subset of rows in the post-hoc tests summary table
generated by \code{makeSummaryTable} which contain only those pairwise treatment
comparisons which differ by exactly one factor level.
}
\examples{
library(predictmeans)
library(nlme)
Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)

# library(lme4)
# fm <- lmer(yield ~ nitro*Variety+(1|Block/Variety), data=Oats)

pm <- predictmeans(fm, "nitro:Variety", pairwise=TRUE)
oats.tab <- makeSummaryTable(pm)

rowsKeep <- getFixedLevelComparisons(oats.tab[,1], sepChar = ",")
oats.tab[rowsKeep,]
}
\seealso{
\code{\link[predictmeans]{predictmeans}}

\code{\link[ruggi]{makeSummaryTable}}

\code{\link[ruggi]{makeComparisonNames}}
}
\author{
Katya Ruggiero
}
