% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLSDsummary.R
\name{getLSDsummary}
\alias{getLSDsummary}
\title{Computes minimum, average and maximum least significant difference (LSD) of selected
pairwise treatment comparisons}
\usage{
getLSDsummary(data, lsdVar, by, alphaName = TRUE)
}
\arguments{
\item{data}{dataframe; see \code{Details} below for requirements and how to generate.}

\item{lsdVar}{character string of column name in \code{data} containing LSDs.}

\item{by}{character vector of factor (column) names in \code{data}. The
minimum, mean and maximum LSD will be computed across all values within each level, or
combination of levels, defined by the factor(s) supplied to this argument.}

\item{alphaName}{remove trailing digits ('1' and '2' only) at the end of \code{by} names.
Default is \code{TRUE}. See \code{Details} for more information.}
}
\value{
A dataframe comprising a column for each factor in \code{by} and a column for the
minimum (min), average (avg) and maximum (max)  LSD for each factorial treatment combination given
by the levels of the factors in \code{by}.
}
\description{
This function uses the LSDs contained in the summary table of pairwise comparisons of means,
generated by \link{makeSummaryTable}, to compute the minimum, average and maximum LSD for
pairwise comparisons of means. See \code{Details}.
}
\details{
The dataframe required by the \code{data} argument in this function may be generated
by the \link{makeSummaryTable} function. If this dataframe is used, the \code{Comparison}
column must be replaced with the dataframe generated by \link{comparisonNames2df}
function, as shown in the examples below. This results in the single \code{Comparison} column
being replaced by \eqn{2 \times n} columns, with these new columns having names ending with either
1 or 2 (see Details section of \link{comparisonNames2df}). The \code{alphaName = TRUE}
argument in the current function drops these trailing digits when naming columns of the
dataframes generated by the current function. This is because the current function is called by
\link{makeLSDlist}, which requires that these column names be identical to those provided
to the \code{by} arg. Thus, default \code{alphaName = TRUE} ensures the matching of these
names.
}
\examples{
library(nlme)
library(predictmeans)
Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)
# library(lme4)
# fm <- lmer(yield ~ nitro*Variety+(1|Block/Variety), data=Oats)
pm <- predictmeans(fm, "nitro:Variety", pairwise=TRUE, plot=FALSE)
pm.tab <- makeSummaryTable(pm)
compNames.df <- comparisonNames2df(pm.tab$Comparison, split.at=":",
                                  varNames=c("nitro", "Variety"))
# keep rows in which Variety1 = Variety2
keepRows <- whichComparison(compNames.df, contrFactor="nitro")
newpm.tab <- data.frame(compNames.df, pm.tab[,-1])[keepRows,]
getLSDsummary(newpm.tab, lsdVar = "LSD", by = "Variety1")
}
\author{
Katya Ruggiero
}
