% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCor.R
\name{groupCor}
\alias{groupCor}
\title{Constructs a table of correlations for each unique combination of factor levels}
\usage{
groupCor(data, measure.vars, factors, useObs = "pairwise.complete.obs",
  useMethod = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{data}{is a dataframe containing both measurement and categorical variables.}

\item{measure.vars}{is a character vector containing the names of the measurement variables in
\code{data} for which pairwise correlations will be calculated.}

\item{factors}{is a character vector containing the names of the categorical variables in
\code{data} from which groups are created.}

\item{useObs}{an optional character string giving a method for computing correlation
in the presence of missing values. This must be (an abbreviation of) one of the strings
\code{"everything"}, \code{"all.obs"}, \code{"complete.obs"}, \code{"na.or.complete"}, or
\code{"pairwise.complete.obs"} (default).}

\item{useMethod}{a character string indicating which correlation coefficient (or covariance)
is to be computed. One of \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"},
can be abbreviated.}
}
\value{
A dataframe containing \eqn{r} rows and \eqn{c} columns, where \eqn{r} equals the product
of the number of levels in each group variable and \eqn{c} is the sum of the number of variables
in \code{factors} and \code{choose(length(measure.vars), 2)}, i.e. the number of pairwise
correlations that can be computed from the set of measurement variables.
}
\description{
This function calculates pairwise correlations between two or more variables by groups, where
groups may be defined as either the levels of a single treatment factor or factorial treatment
combinations.
}
\examples{
data(fruitfly)
groupCor(fruitfly, factors=c("Type"), measure.vars=c("Longevity", "Thorax"))
groupCor(fruitfly, factors=c("Partners", "Type"), measure.vars=c("Longevity", "Thorax", "Sleep"))
}
