% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComparisonNames.R
\name{makeComparisonNames}
\alias{makeComparisonNames}
\title{Creates pairwise treatment comparison names}
\usage{
makeComparisonNames(mat)
}
\arguments{
\item{mat}{is any square symmetric matrix with row and column names being treatment (or
factorial treatment combination) names. See \code{Details} below.}
}
\value{
a character vector of all pairwise treatment comparisons names
}
\description{
This function creates a character vector of pairwise treatment comparison names using the
dimension names of a matrix object such as the \code{Pairwise LSDs} and \code{Pairwise p-value}
matrices generated by the \code{predictmeans()} function in the \code{predictmeans} package.
}
\details{
Although a standalone function, it was primarily developed for its call by the
\link{makeSummaryTable} function to create a column of character strings defining the pair of
treatments being contrasted (i.e. the pairwise comparisons of means) in the summary table.
In \link{makeSummaryTable}, it uses the dimension names of the \code{Pairwise p-value} matrix
(generated by \link[predictmeans]{predictmeans}) to construct a character string comprising the
pair of treatment names (for each pairwise comparison of means) concatenated with a minus sign
'-' in the middle.

Note that this function also works with the p-value matrix generated by the
\code{predictparallel.asreml()} function in the \code{asremlPlus} package.
}
\examples{
library(ggplot2)
library(nlme)
library(predictmeans)

Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)

library(lme4)
fm <- lmer(yield ~ nitro*Variety+(1|Block/Variety), data=Oats)
pm <- predictmeans(fm, "nitro:Variety", pairwise=TRUE)
makeComparisonNames(pm$"Pairwise p-values")
}
\seealso{
\code{\link[predictmeans]{predictmeans}}
}
\author{
Katya Ruggiero
}
