% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummaryTable.R
\name{makeSummaryTable}
\alias{makeSummaryTable}
\title{Constructs a table of summary statistics for post-hoc tests of pairwise differences between means
from a linear (mixed) model}
\usage{
makeSummaryTable(pm, model.term, wald.tab, alpha, digits = c(3, 4),
  eps = NULL)
}
\arguments{
\item{pm}{is a list generated by the \code{predictmeans} function. \code{predictmeans} takes as
its first argument the results of a fitted model using any of \code{lm}, \code{aov},
\code{glm}, \code{lme} or \code{lmer}.}

\item{model.term}{(only required if input table of predicted values is generated by
\code{predictparallel.asreml} function in the \code{asremlPlus} package) name, in "quotes",
of main effect or interaction (e.g. \code{A}, \code{A:B}) for which predicted means were
 calculated by either \code{predictmeans} or \code{asremlPlus}. \code{modelterm} must be
 given exactly as it appears in the ANOVA table.}

\item{wald.tab}{(only required if input table of predicted values is generated by
\code{predictparallel.asreml} function in the \code{asremlPlus} package) table of incremental
or conditional F statistics generated by \code{asreml}'s \code{wald} function.}

\item{alpha}{(only required if input table of predicted values is generated by
\code{predictparallel.asreml} function in the \code{asremlPlus} package) level of significance
at which to perform post-hoc tests.}

\item{digits}{integer vector of length 2 indicating the number of decimal places to be used
for the output values of \code{Difference} -- \code{upr} (first integer in vector, default 3)
and for the \code{p}-values (second integer in vector, default 4).}

\item{eps}{user-defined tolerance level (default \code{NULL}) for dealing with floating-point
errors which erroneously yield non-zero differences between pairs of means.}
}
\value{
A dataframe containing:

\code{Comparison} character string of the names of the two groups' means being compared

\code{Difference} value the difference between the pair of means

\code{SED} standard error of the difference between the pair of means

\code{t} t-statistic

\code{LSD} least significant difference calculated at the \eqn{\alpha}
level of significance. The value of \eqn{\alpha} is defined in the \code{predictmeans} function

\code{lwr} lower 100 \eqn{\times (1-\alpha)}\% confidence limit of the difference

\code{upr} upper 95\% confidence limit of the difference

\code{p} p-value
}
\description{
This function takes the contents of either the list generated by the \code{predictmeans} function
in the \code{predictmeans} package or the list generated by \code{predictparallel.asreml}
function in the \code{asremlPlus} package to generate a dataframe which summarises the results
of all pairwise comparisons of means for the \code{modelterm} provided to the \code{predictmeans}
function.
}
\details{
Note: \code{predictparallel.asreml} requires an object of class \code{asreml} containing
the results of a fitted mixed model generated by the \code{asreml} function in the
commercial \code{R} package \code{asreml}.

This function generates a dataframe with as many rows as there are pairwise comparisons of means
and columns containg the comparison description, the difference between the pair of means,
the standard error of the difference, the least significant difference, the t-statistics, lower
and upper pairwise confidence limits and p-value.
}
\examples{
library(predictmeans)
library(nlme)
Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)
# library(lme4)
# fm <- lmer(yield ~ nitro*Variety+(1|Block/Variety), data=Oats)
pm <- predictmeans(fm, "nitro:Variety", pairwise=TRUE, plot=FALSE)
makeSummaryTable(pm)

# The following example requires the commercial \\code{R} package \\code{asreml}
# to run.
# library(asreml)
# library(asremlPlus)
# oats.asr <- asreml(yield ~ nitro*Variety, random = ~Block/Variety, data=Oats)
# oats.wld <- wald(oats.asr, denDF = "default", ssType = "conditional", maxiter = 1)$Wald
# oats.ppasr <- predictparallel.asreml(classify = "nitro:Variety",
#    asreml.obj=oats.asr, wald.tab = oats.wld,
#    tables = "none")
# oats.tab <- makeSummaryTable(oats.ppasr, model.term = "nitro:Variety",
#    wald.tab = oats.wld, alpha = 0.05)

}
\seealso{
\code{\link[predictmeans]{predictmeans}}, \link{makeComparisonNames}
}
\author{
Katya Ruggiero

Kathy Ruggiero
}
