% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanse.R
\name{meanse}
\alias{meanse}
\title{Calculate mean and standard error}
\usage{
meanse(x)
}
\arguments{
\item{x}{is a vector of observations}
}
\value{
a vector of two elements, the mean and standard error of observations in x
}
\description{
This function calculates the mean and standard error of the data in a vector 'x'.
}
\details{
This function computes the mean and standard error of the observations in a vector.
NAs are removed before the calculations are performed.
}
\seealso{
\code{\link{mean}}, \code{\link{sd}}
}
\author{
Katya Ruggiero
}
