% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdData.R
\docType{data}
\name{rcbdData}
\alias{rcbdData}
\title{Simulated data set for a randomised complete block design (RCBD)}
\format{A data frame with 12 rows and 5 variables:
\describe{
  \item{Run}{block factor corresponding to independent runs of the mass spectrometer}
  \item{Surgery}{treatment factor corresponding type of surgical conditioning}
  \item{Rat}{animal IDs}
  \item{logAUC8}{log-transformed area under the curve, a proxy for log-abundance of protein.
  The data was simulated assuming a between run error variance of 8 and a within run error
  variance of 1.}
  \item{logAUC4}{log-transformed area under the curve, a proxy for log-abundance of protein.
  The data was simulated assuming a between run error variance of 4 and a within run error
  variance of 1.}
}}
\usage{
rcbdData
}
\description{
A dataset containing the log-transformed protein abundances
in 12 rats randomly assigned to one of 3 surgical conditionings.
The rat samples are allocated to 4 runs, each comprising the
three treatment groups.
}
\keyword{datasets}
