% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBlockData.R
\docType{data}
\name{splitBlockData}
\alias{splitBlockData}
\title{Simulated data set for a split-block design}
\format{A data frame with 12 rows and 5 variables:
\describe{
  \item{Disease}{disease status, either Healthy or Diabetic}
  \item{Organ}{harvested tissue, namely the inner and outer LV}
  \item{Animal}{Rat ID}
  \item{Sample}{an arbitrary label (either 1 or 2) to different the two samples taken from
  each rat.}
  \item{logAUC}{log-transformed area under the curve, a proxy for log-abundance of protein.}
}}
\usage{
splitBlockData
}
\description{
A dataset containing the log-transformed protein abundances
in 6 rats randomly assigned to one of 2 treatment groups:
Healthy or Diabetic. Two organs were harvested from each animal,
namely the inner and outer left ventricle (LV) wall.
}
\keyword{datasets}
