% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichComparison.R
\name{whichComparison}
\alias{whichComparison}
\title{Generates a logical vector for subsetting from a table of pairwise comparisons of treatments}
\usage{
whichComparison(data, contrFactor)
}
\arguments{
\item{data}{a dataframe containing 2n columns, where n is the number of factors contributing to
a factorial treatment combination. The labels in the first n columns contain the levels
contributing
to the factorial treatment combination of the first mean in the comparison, while the second n
columns contain the levels contribution to the second mean.}

\item{contrFactor}{name of factor with different values (levels) in the two factorial treatment
combinations being contrasted.}
}
\value{
a vector of logicals which has as many elements as there are rows in \code{data}.
}
\description{
The vector of logicals generated by this function yields a TRUE value for each pairwise
comparison of factorial treatment combinations in which all, but one, of the matching factors
in the two treatment combinations have the same value (level). This function requires a
dataframe in which the character string for each factorial treatment combination has been
decoupled into its original components, i.e. a column per factor making up each factorial
treatment combination. See \code{Details} for more information.
}
\details{
The dataframe required by the \code{data} argument in this function may
be generated by applying the
\link{comparisonNames2df} function to the \code{Comparison} column in the dataframe
generated by the \link{makeSummaryTable} function available in this package.
}
\examples{
library(predictmeans)
library(nlme)
Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)

# library(lme4)
# fm <- lmer(yield ~ nitro*Variety+(1|Block/Variety), data=Oats)

pm <- predictmeans(fm, "nitro:Variety", pairwise=TRUE, plot=FALSE)
oats.tab <- makeSummaryTable(pm)
compNames.df <- comparisonNames2df(oats.tab[,1], split.at = ":", varNames = c("nitro", "Variety") )
newoats.tab <- data.frame(compNames.df, oats.tab[,-1])

# pairwise comparisons between nitro levels within varieties
keeprows <- whichComparison(compNames.df, contrFactor="nitro")
newoats.tab[keeprows,]
}
\seealso{
\code{\link[ruggi]{makeComparisonNames}}

\code{\link[ruggi]{makeSummaryTable}}

\code{\link[ruggi]{comparisonNames2df}}
}
\author{
Katya Ruggiero
}
